Core: Updated to 11.1.7
Port From (https://github.com/TrinityCore/TrinityCore/commit/ceb8d561b44579118b109a9ff71ba415df53bbdc)

+++ Source/Game/DataStorage/DB2Manager.cs
@@ -1944,7 +1944,7 @@ namespace Game.DataStorage
             if (v2 == null)
                 return 0;
 
-            return v2.UniqueBitFlag;
+            return (uint)v2.UniqueBitFlag;
         }
 
         public List<uint> GetPhasesForGroup(uint group)
